pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "simulation_based_graph_inference"
version = "0.1"
dependencies = [
    "collectiontools>=0.2.0",
    "fasttr",
    "matplotlib>=3.10.6",
    "scipy>=1.16.2",
    "torch==2.8.*",
    "torch-geometric>=2.6.1",
    "torch-scatter>=2.1.2",
]
requires-python = ">=3.12"

[dependency-groups]
dev = [
    "cook-build>=0.7.4",
    "jupyter>=1.1.1",
    "jupytext>=1.17.3",
    "line-profiler>=5.0.2",
    "localscope>=0.2.5",
    "pre-commit>=4.5.1",
    "pyright>=1.1.407",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.13.1",
    "sphinx>=8.2.3",
]

[tool.uv]
find-links = [
    "https://data.pyg.org/whl/torch-2.8.0+cpu.html",
]

[tool.uv.extra-build-dependencies]
torch-scatter = ["torch"]
fasttr = ["pybind11"]

[tool.uv.sources]
fasttr = { git = "https://github.com/gstonge/fasttr" }
